requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Werkzeug
WTForms
python-dotenv
